Use app/groonga instead of app/models
