+++ app/groonga/post_search_result_set.rb
