+++ app/groonga/query.rb
